refactor: migrate help group
