+++ packages/webpack-cli/lib/utils/unknown-args.js
@@ -5,6 +5,6 @@ module.exports = {
     commands: [...commandNames],
     flags: [...flagNames],
     allNames: [...commandNames, ...flagNames],
-    hasUnknownArgs: (args, ...names) =>
+    hasUnknownArgs: (args, names) =>
         args.filter((e) => !names.includes(e) && !e.includes('color') && e !== 'version' && e !== '-v' && !e.includes('help')),
 };
